rework `InnerHandler` into daemon state
