+++ src/lib.rs
@@ -5,3 +5,4 @@ pub mod shutting_down_stream;
 pub mod input_token;
 pub mod byte_slices;
 pub mod commands;
+mod daemon_state;
